Fix benchmark speed reporting of very fast formats (eg. LM-opencl w/ mask).  This bug was introduced at core merge 844df7388 because the replaced int64 "add32to64()" function assumed unsigned whereas the new code did not.

+++ src/bench.c
@@ -514,7 +514,7 @@ char *benchmark_format(struct fmt_main *format, int salts,
 		    format->methods.crypt_all(&count, 0));
 #endif
 
-		crypts += count;
+		crypts += (uint32_t)count;
 #if !OS_TIMER
 		sig_timer_emu_tick();
 #endif
